Move distance calculations up
