+++ markers/+components/PositionBar.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { Platform, StyleSheet, View, Animated } from "react-native";
 import { StyledText } from "../../+components/StyledText";
-import { deg2rad } from "../+utils/distance";
+import { deg2rad } from "../../+utils/distanceCalculator";
 
 export const PositionBar = ({ location, isHidden }) => {
   const [hideAnimation] = useState(new Animated.Value(0));
